% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emy_data.R
\docType{data}
\name{emy_data}
\alias{emy_data}
\title{Climatic data from National Meteorological Service}
\format{A data frame with 1679308 rows and 13 variables:
\describe{
  \item{date}{the observation's date}
  \item{station_id}{Greek National Meteorological Service station ID}
  \item{pressure}{Pressure (hPa)}
  \item{dry_temp}{Dry temperature (degrees C)}
  \item{wet_temp}{Wet temperature (degrees C)}
  \item{max_temp}{Maximum temperature (degrees C)}
  \item{min_temp}{Minimum temperature (degrees C)}
  \item{rel_humidity}{Relative humidity (\%)}
  \item{wind_direction}{Wind direction (degrees)}
  \item{wind_force_knot}{Wind force (knots)}
  \item{visibility}{Visibility (km)}
  \item{total_cloud_amount}{Total cloud amount (-)}
  \item{precipitation}{Precipitation height (mm)}
}}
\source{
\url{http://www.emy.gr/}
}
\usage{
emy_data
}
\description{
The climatic data are taken from the Greek National Meteorological Service
and come from 21 weather stations, with a time step of 3 hours. The time
period is from 1975 to 2004. These time series are not checked for
consistency or errors.
}
\keyword{datasets}
