% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emy_stations}
\alias{emy_stations}
\title{Greek National Meteorological Service Stations}
\format{A data frame with 21 rows and 7 variables:
\describe{
  \item{name}{The station's name}
  \item{station_id}{Greek National Meteorological Service station ID}
  \item{hydroscope_id}{The station's ID from kyy.hydroscope.gr database}
  \item{water_division}{The station's Water Division}
  \item{longitude}{The station's longitude in decimal degrees, ETRS89}
  \item{latitude}{The station's latitude in decimal degrees, ETRS89}
  \item{altitude}{The station's altitude, meters above sea level}
}}
\source{
\url{http://www.hydroscope.gr/}
}
\usage{
emy_stations
}
\description{
The stations' data are taken from the Greek National Data Bank for
Hydrological and Meteorological Information, Hydroscope. This data frame
has the stations' id, coordinates, altitude and water division.
}
\keyword{datasets}
