% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emyweather-package.R
\docType{package}
\name{emyweather}
\alias{emyweather}
\alias{_PACKAGE}
\alias{emyweather-package}
\alias{emyweather-package}
\title{emyweather: Open data from the Greek National Meteorological Service}
\description{
This package contains climatic data from the Greek National
Meteorological Service, as well as a number of useful other related data
sets.
}
\section{Climatic Data}{


The climatic data are taken from the Greek National Meteorological Service
and come from 21 weather stations, with a time step of 3 hours. The time
period is from 1975 to 2004. The variables are: air pressure, temperature
(dry, wet, max and min), relative humidity, wind (direction and speed),
visibility, total cloud amount and precipitation. These time series are not
checked for consistency or errors.
}

\section{Data Sources}{


The data are retrieved from http://www.meteo.gr/
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvantas/emyweather}
  \item Report bugs at \url{https://github.com/kvantas/emyweather/issues}
}

}
\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com} (0000-0001-6387-8791)

}
